separate test modules + add validator to game.Player.Action init
